Rename external.rs to testutils.rs
