+++ src/lib.rs
@@ -9,8 +9,8 @@ mod allowance;
 mod balance;
 mod contract;
 mod cryptography;
-pub mod external;
 mod metadata;
 mod nonce;
 pub mod public_types;
 mod storage_types;
+pub mod testutils;
